requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==22.0.0
werkzeug==3.1.3
boto3>=1.35.0
Pillow>=10.0.0
requests>=2.31.0
anthropic>=0.40.0
